First clean
Renamed the top and top2 to Top-first-row and top-second-row. Removed an unmotivated space in section center-text.

In Css i removed outcommeted Styles. I made indentation and added space between classes where it was missing. I also added twiligt mode to smaller screens, was only created for laptop views.
I also moved around code so that it made sense depeding on where the html is on the page, so the mountains is in the bottom and the loading was moved up above, and move down "nightsky" so that its not in the begining of the styling.