Added ; and removed/added space (also rearanged some functions)
added ; to all the functions so that it was coheasive thoughout the document, some had some where missing. There was also some spaces here and there that i added or removed. I also rearanged som functions so that they are now structured a bit from "top of page to bottom" in the changes. Biggest change was nightmode