fixed twilight advice issue
The twiligt advice click funktion was not right since only the sun was moved down, and not the clickable area who listening for clicks. So changed so that the area move too and not just the sun. Then i moved it ever future down becasue it was not clickable because it was behind the counter box, so like only the top of the sun/moon was clickable.