Removed cors and simplify checkTime function
The cors shortcut that we used had stopped being free and stoped working. Removed that from the URL and also simplified the checkTime to only check if its right before sunset or after sunset, then "else" do daytime. There was something that was not working correctly in the if, so it just checked out and nothing was showing